refactor: move stats DTOs to metrics folder, and add maven deps for micrometer metrics instruments
