+++ project/src/main/java/org/ucd/shortlink/project/controller/ShortLinkStatsController.java
@@ -24,10 +24,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.ucd.shortlink.project.common.convention.result.Result;
 import org.ucd.shortlink.project.common.convention.result.Results;
-import org.ucd.shortlink.project.dto.req.ShortLinkGroupStatsAccessRecordReqDTO;
-import org.ucd.shortlink.project.dto.req.ShortLinkStatsReqDTO;
-import org.ucd.shortlink.project.dto.resp.ShortLinkStatsAccessRecordRespDTO;
-import org.ucd.shortlink.project.dto.resp.ShortLinkStatsRespDTO;
+import org.ucd.shortlink.project.dto.req.metrics.ShortLinkGroupStatsAccessRecordReqDTO;
+import org.ucd.shortlink.project.dto.req.metrics.ShortLinkStatsReqDTO;
+import org.ucd.shortlink.project.dto.resp.metrics.ShortLinkStatsAccessRecordRespDTO;
+import org.ucd.shortlink.project.dto.resp.metrics.ShortLinkStatsRespDTO;
 import org.ucd.shortlink.project.service.ShortLinkStatsService;
 
 /**
