+++ project/src/main/java/org/ucd/shortlink/project/dao/mapper/LinkBrowserStatsMapper.java
@@ -22,8 +22,8 @@ import org.apache.ibatis.annotations.Insert;
 import org.apache.ibatis.annotations.Param;
 import org.apache.ibatis.annotations.Select;
 import org.ucd.shortlink.project.dao.entity.LinkBrowserStatsDO;
-import org.ucd.shortlink.project.dto.req.ShortLinkGroupStatsReqDTO;
-import org.ucd.shortlink.project.dto.req.ShortLinkStatsReqDTO;
+import org.ucd.shortlink.project.dto.req.metrics.ShortLinkGroupStatsReqDTO;
+import org.ucd.shortlink.project.dto.req.metrics.ShortLinkStatsReqDTO;
 
 import java.util.HashMap;
 import java.util.List;
