+++ project/src/main/java/org/ucd/shortlink/project/dto/resp/metrics/ShortLinkStatsTopIpRespDTO.java
@@ -15,62 +15,28 @@
  * limitations under the License.
  */
 
-package org.ucd.shortlink.project.dto.resp;
+package org.ucd.shortlink.project.dto.resp.metrics;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.Date;
-
 /**
- * Short link monitor metric record response bod
+ * High frequency access IP monitor state response body
  */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ShortLinkStatsAccessRecordRespDTO {
+public class ShortLinkStatsTopIpRespDTO {
     /**
-     * Unique visitor type
+     * Total cnt
      */
-    private String uvType;
-
-    /**
-     * Browser
-     */
-    private String browser;
-
-    /**
-     * OS type
-     */
-    private String os;
+    private Integer cnt;
 
     /**
      * IP
      */
     private String ip;
-
-    /**
-     * Device type
-     */
-    private String device;
-
-    /**
-     * Region
-     */
-    private String locale;
-
-    /**
-     * User info
-     */
-    private String user;
-
-    /**
-     * Visit timestamp
-     */
-    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
-    private Date createTime;
 }
