+++ project/src/main/java/org/ucd/shortlink/project/service/ShortLinkStatsPromServiceImpl.java
@@ -20,11 +20,11 @@ package org.ucd.shortlink.project.service;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.ucd.shortlink.project.dto.req.ShortLinkGroupStatsAccessRecordReqDTO;
-import org.ucd.shortlink.project.dto.req.ShortLinkStatsAccessRecordReqDTO;
-import org.ucd.shortlink.project.dto.req.ShortLinkStatsReqDTO;
-import org.ucd.shortlink.project.dto.resp.ShortLinkStatsAccessRecordRespDTO;
-import org.ucd.shortlink.project.dto.resp.ShortLinkStatsRespDTO;
+import org.ucd.shortlink.project.dto.req.metrics.ShortLinkGroupStatsAccessRecordReqDTO;
+import org.ucd.shortlink.project.dto.req.metrics.ShortLinkStatsAccessRecordReqDTO;
+import org.ucd.shortlink.project.dto.req.metrics.ShortLinkStatsReqDTO;
+import org.ucd.shortlink.project.dto.resp.metrics.ShortLinkStatsAccessRecordRespDTO;
+import org.ucd.shortlink.project.dto.resp.metrics.ShortLinkStatsRespDTO;
 import org.ucd.shortlink.project.prometheus.service.PrometheusService;
 
 /**
